+++ src/config.js
@@ -1,4 +1,3 @@
-
 const config = {
   /** CLI flag */
   generateResponses: false,
@@ -8,14 +7,17 @@ const config = {
   generateRouteTypes: false,
   /** CLI flag */
   generateClient: true,
-  /** parsed swagger schema from getSwaggerObject() */ 
+  /** parsed swagger schema from getSwaggerObject() */
+
   swaggerSchema: null,
   /** { "#/components/schemas/Foo": @TypeInfo, ... } */
   componentsMap: {},
-}
+  /** flag for catching convertion from swagger 2.0 */
+  convertedFromSwagger2: false,
+};
 
 /** needs to use data everywhere in project */
 module.exports = {
-  addToConfig: configParts => Object.assign(config, configParts),
+  addToConfig: (configParts) => Object.assign(config, configParts),
   config,
-} 
+};